feat: Minor changes for env replacement
