+++ packages/amplify-cli-core/src/customPoliciesUtils.ts
@@ -15,3 +15,7 @@ export const CustomIAMPolicySchema = {
         required: ["Resource", "Action"],
         additionalProperties: false
 }
+
+
+
+
